requirements.txt
pytest ==  8.4.0
pytest-playwright==0.7.0
pytest-xdist==3.7.0
pytest-html==4.1.1